backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rashad Philizaire <rashadphilizaire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.2"}
pydantic = "^2.7.1"
requests = "^2.31.0"
httpx = "^0.27.0"
starlette = "^0.37.2"
sse-starlette = "^2.1.0"
tavily-python = "^0.3.3"
load-dotenv = "^0.1.0"
llama-index = "^0.10.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
